+++ CHANGES.md
@@ -8,6 +8,7 @@
 - macOS: Native support for Apple Silicon macs
 - iOS: Fix 'Change Location' to make first location choosable #375
 - Use strings in discovery data correctly based on system language #374
+- Auto-focus the search field when applicable #330
 
 ## 2.2.0 (2020-11-12)